LayoutTests/http/wpt/webauthn/public-key-credential-get-success-hid.https.html

<!DOCTYPE html><!-- webkit-test-runner [ WebAuthenticationModernEnabled=false ] -->
<title>Web Authentication API: PublicKeyCredential's [[get]] success cases with a mock hid authenticator.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/util.js"></script>
<script>
    // Default mock configuration. Tests need to override if they need different configuration.
    if (window.internals)
        internals.setMockWebAuthenticationConfiguration({ hid: { stage: "request", subStage: "msg", error: "success", payloadBase64: [testAssertionMessageBase64] } });

    promise_test(t => {
        const options = {
            publicKey: {
                challenge: Base64URL.parse("MTIzNDU2"),
                timeout: 100
            }
        };

        let expected = {
            "authenticatorAttachment" : "cross-platform",
            "clientExtensionResults":{},
            "id":"KAitzuj+Tslzelf3/vZwIGtDQNgoKeFd5oEieYzhyzA65saf0tK2w/mooa7tQtGgDdwZIjOhjcuZ0pQ1ajoE4A==",
            "rawId":"KAitzuj+Tslzelf3/vZwIGtDQNgoKeFd5oEieYzhyzA65saf0tK2w/mooa7tQtGgDdwZIjOhjcuZ0pQ1ajoE4A==",
            "response" : {
                "authenticatorData":"Rsx/uWedVbLbkJLhyNnl4dArdYDwtIEsdwli4eSPWtgBAAAAUA==",
                "clientDataJSON":"eyJ0eXBlIjoid2ViYXV0aG4uZ2V0IiwiY2hhbGxlbmdlIjoiTVRJek5EVTIiLCJvcmlnaW4iOiJodHRwczovL2xvY2FsaG9zdDo5NDQzIn0=",
                "signature":"MEUCIQCSFTuuBWgB4/F0VB7DlUVM09IHPmxe1MzHUwRoCRZbCAIgGKov6xoAx2MEf6/6qNs8OutzhP2CQoJ1L7Fe64G9uBc="
            },
            "type":"public-key"
        };

        return navigator.credentials.get(options).then(credential => {
            assert_equals(JSON.stringify(expected), JSON.stringify(credential.toJSON()));
            return checkCtapGetAssertionResult(credential);
        });
    }, "PublicKeyCredential's [[get]] calling toJSON on the result.");

    promise_test(t => {
        const options = {
            publicKey: {
                challenge: Base64URL.parse("MTIzNDU2"),
                timeout: 100
            }
        };

        return navigator.credentials.get(options).then(credential => {
            return checkCtapGetAssertionResult(credential);
        });
    }, "PublicKeyCredential's [[get]] with minimum options in a mock hid authenticator.");

    promise_test(t => {
        const options = {
            publicKey: {
                challenge: Base64URL.parse("MTIzNDU2"),
                allowCredentials: [
                    { type: "public-key", id: Base64URL.parse(testHidCredentialIdBase64), transports: ["usb"] }
                ],
                timeout: 100
            }
        };

        return navigator.credentials.get(options).then(credential => {
            return checkCtapGetAssertionResult(credential);
        });
    }, "PublicKeyCredential's [[get]] with matched allow credentials in a mock hid authenticator.");

    promise_test(t => {
        const options = {
            publicKey: {
                challenge: Base64URL.parse("MTIzNDU2"),
                userVerification: "preferred",
                timeout: 100
            }
        };

        return navigator.credentials.get(options).then(credential => {
            return checkCtapGetAssertionResult(credential);
        });
    }, "PublicKeyCredential's [[get]] with userVerification { preferred } in a mock hid authenticator.");

    promise_test(t => {
        const options = {
            publicKey: {
                challenge: Base64URL.parse("MTIzNDU2"),
                userVerification: "discouraged",
                timeout: 100
            }
        };

        return navigator.credentials.get(options).then(credential => {
            return checkCtapGetAssertionResult(credential);
        });
    }, "PublicKeyCredential's [[get]] with userVerification { discouraged } in a mock hid authenticator.");

    promise_test(t => {
        const options = {
            publicKey: {
                challenge: Base64URL.parse("MTIzNDU2"),
                allowCredentials: [
                    { type: "public-key", id: Base64URL.parse(testHidCredentialIdBase64), transports: ["usb"] }
                ],
                userVerification: "preferred",
                timeout: 100
            }
        };

        return navigator.credentials.get(options).then(credential => {
            return checkCtapGetAssertionResult(credential);
        });
    }, "PublicKeyCredential's [[get]] with mixed options in a mock hid authenticator.");

    promise_test(t => {
        const options = {
            publicKey: {
                challenge: Base64URL.parse("MTIzNDU2"),
                timeout: 100
            }
        };


        // Stall the first request to wait for cancellation.
        if (window.internals)
            internals.setMockWebAuthenticationConfiguration({ hid: { stage: "request", subStage: "msg", error: "success", payloadBase64: [testAssertionMessageBase64], expectCancel: true } });
        promiseRejects(t, "NotAllowedError", navigator.credentials.get(options), "This request has been cancelled by a new request.");
        if (window.internals)
            internals.setMockWebAuthenticationConfiguration({ hid: { stage: "request", subStage: "msg", error: "success", payloadBase64: [testAssertionMessageBase64] } });
        return navigator.credentials.get(options).then(credential => {
            return checkCtapGetAssertionResult(credential);
        });
    }, "PublicKeyCredential's [[get]] with two consecutive requests.");

    promise_test(t => {
        const options = {
            publicKey: {
                challenge: Base64URL.parse("MTIzNDU2"),
                timeout: 100
            }
        };

        if (window.internals)
            internals.setMockWebAuthenticationConfiguration({ hid: { stage: "request", subStage: "msg", error: "success", payloadBase64: [testAssertionMessageLongBase64, testAssertionMessageLongBase64] } });
        return navigator.credentials.get(options).then(credential => {
            return checkCtapGetAssertionResult(credential, "MIIBkzCCATigAwIBAjCCAZMwggE4oAMCAQIwggGTMII=");
        });
    }, "PublicKeyCredential's [[get]] with multiple accounts in a mock hid authenticator.");

    promise_test(t => {
        const options = {
            publicKey: {
                challenge: Base64URL.parse("MTIzNDU2"),
                userVerification: "discouraged",
                timeout: 100
            }
        };

        if (window.internals)
            internals.setMockWebAuthenticationConfiguration({ hid: { supportClientPin: true, payloadBase64: [testAssertionMessageBase64] } });
        return navigator.credentials.get(options).then(credential => {
            return checkCtapGetAssertionResult(credential);
        });
    }, "PublicKeyCredential's [[get]] with PIN supported in the authenticator but userVerification = 'discouraged' in a mock hid authenticator.");
</script>
